+++ libs/components/src/lib/components/dropdowns/dropdown-host/models.ts
@@ -1,7 +1,20 @@
-import { PrizmDropdownZoneDirective } from '../../../directives/event-zone/event-zone.directive';
-
 export type PrizmDropdownHostWidth = string | number | null;
 export type PrizmDropdownHostContext = {
   custom: PrizmDropdownHostCustomContext;
 };
 export type PrizmDropdownHostCustomContext = Record<string, unknown>;
+export type PrizmDropdownHostStyles =
+  | {
+      [klass: string]: any;
+    }
+  | null
+  | undefined;
+export type PrizmDropdownHostClasses =
+  | string
+  | string[]
+  | Set<string>
+  | {
+      [klass: string]: boolean;
+    }
+  | null
+  | undefined;